# openapi.json
{"openapi":"3.1.0","info":{"title":"Arbitrage Sourcing Engine","description":"Retail-arbitrage lead engine. Ingests retailer catalogs, detects active sales, and computes profit / ROI / margin.\n\n**Amazon-side data is modelled, not live.** Every lead carries `modelled: true` until a Keepa API key is configured. Do not present modelled ROI as a verified sourcing decision.","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["meta"],"summary":"Health","description":"Liveness plus an honest statement of what the Amazon side can do.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Health"}}}}}}},"/stats":{"get":{"tags":["meta"],"summary":"Stats","operationId":"stats_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Stats"}}}}}}},"/retailers":{"get":{"tags":["retailers"],"summary":"List Retailers","operationId":"list_retailers_retailers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Retailer"},"type":"array","title":"Response List Retailers Retailers Get"}}}}}}},"/sales":{"get":{"tags":["leads"],"summary":"List Sales","description":"Products currently discounted and in stock. Real retailer data.","operationId":"list_sales_sales_get","parameters":[{"name":"min_discount","in":"query","required":false,"schema":{"type":"number","maximum":100,"minimum":0,"default":15.0,"title":"Min Discount"}},{"name":"retailer","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retailer"}},{"name":"min_price","in":"query","required":false,"schema":{"type":"number","minimum":0,"default":0.5,"title":"Min Price"}},{"name":"max_price","in":"query","required":false,"schema":{"anyOf":[{"type":"number","exclusiveMinimum":0},{"type":"null"}],"title":"Max Price"}},{"name":"dedup","in":"query","required":false,"schema":{"type":"boolean","description":"Collapse colourway/size variants","default":true,"title":"Dedup"},"description":"Collapse colourway/size variants"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalePage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/leads":{"get":{"tags":["leads"],"summary":"List Leads","description":"Profit-modelled leads.\n\nRetailer cost and discount are real. Amazon price, net profit and ROI are\nMODELLED until a Keepa key exists - see the `modelled` flag.","operationId":"list_leads_leads_get","parameters":[{"name":"min_roi","in":"query","required":false,"schema":{"type":"number","description":"Minimum ROI percent","default":30.0,"title":"Min Roi"},"description":"Minimum ROI percent"},{"name":"multiplier","in":"query","required":false,"schema":{"type":"number","maximum":5,"exclusiveMinimum":0,"description":"Modelled Amazon price as a fraction of list","default":0.85,"title":"Multiplier"},"description":"Modelled Amazon price as a fraction of list"},{"name":"min_discount","in":"query","required":false,"schema":{"type":"number","maximum":100,"minimum":0,"default":15.0,"title":"Min Discount"}},{"name":"retailer","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retailer"}},{"name":"min_price","in":"query","required":false,"schema":{"type":"number","minimum":0,"default":5.0,"title":"Min Price"}},{"name":"max_price","in":"query","required":false,"schema":{"type":"number","exclusiveMinimum":0,"default":200.0,"title":"Max Price"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeadPage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/candidates":{"get":{"tags":["leads"],"summary":"Funnel","description":"How many Keepa lookups a full catalog actually requires.\n\nFiltering before spending a token is what makes hundreds of retailers viable.","operationId":"funnel_candidates_get","parameters":[{"name":"min_price","in":"query","required":false,"schema":{"type":"number","default":5.0,"title":"Min Price"}},{"name":"max_price","in":"query","required":false,"schema":{"type":"number","default":200.0,"title":"Max Price"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Funnel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/export.csv":{"get":{"tags":["leads"],"summary":"Export Csv","description":"Discounted products as CSV.","operationId":"export_csv_export_csv_get","parameters":[{"name":"min_discount","in":"query","required":false,"schema":{"type":"number","default":15.0,"title":"Min Discount"}},{"name":"retailer","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retailer"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":10000,"minimum":1,"default":1000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/retailers/{slug}/ingest":{"post":{"tags":["retailers"],"summary":"Run Ingest","description":"Pull a retailer's catalog and record any price changes.\n\nSynchronous by design at phase-1 scale. Move to a queue before adding\nretailers with large catalogs.","operationId":"run_ingest_retailers__slug__ingest_post","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}},{"name":"pages","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":200,"minimum":1},{"type":"null"}],"title":"Pages"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/fingerprint":{"get":{"tags":["retailers"],"summary":"Fingerprint","description":"Sort a domain into an acquisition tier before committing to it.","operationId":"fingerprint_fingerprint_get","parameters":[{"name":"host","in":"query","required":true,"schema":{"type":"string","description":"Domain, e.g. www.vitacost.com","title":"Host"},"description":"Domain, e.g. www.vitacost.com"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Funnel":{"properties":{"skus_ingested":{"type":"integer","title":"Skus Ingested"},"discounted_in_stock":{"type":"integer","title":"Discounted In Stock"},"discounted_pct":{"type":"number","title":"Discounted Pct"},"within_price_band":{"type":"integer","title":"Within Price Band"},"band_pct":{"type":"number","title":"Band Pct"},"keepa_lookups_needed":{"type":"integer","title":"Keepa Lookups Needed"},"lookup_pct":{"type":"number","title":"Lookup Pct"},"reduction_factor":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Reduction Factor"}},"type":"object","required":["skus_ingested","discounted_in_stock","discounted_pct","within_price_band","band_pct","keepa_lookups_needed","lookup_pct","reduction_factor"],"title":"Funnel"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Health":{"properties":{"status":{"type":"string","title":"Status"},"keepa_configured":{"type":"boolean","title":"Keepa Configured"},"amazon_data":{"type":"string","title":"Amazon Data"}},"type":"object","required":["status","keepa_configured","amazon_data"],"title":"Health"},"IngestResult":{"properties":{"retailer":{"type":"string","title":"Retailer"},"seen":{"type":"integer","title":"Seen"},"new":{"type":"integer","title":"New"},"price_changes":{"type":"integer","title":"Price Changes"},"on_sale":{"type":"integer","title":"On Sale"}},"type":"object","required":["retailer","seen","new","price_changes","on_sale"],"title":"IngestResult"},"Lead":{"properties":{"retailer":{"type":"string","title":"Retailer"},"retailer_slug":{"type":"string","title":"Retailer Slug"},"product_id":{"type":"integer","title":"Product Id"},"title":{"type":"string","title":"Title"},"brand":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"pack_qty":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pack Qty"},"price":{"type":"number","title":"Price"},"list_price":{"type":"number","title":"List Price"},"discount_pct":{"type":"number","title":"Discount Pct"},"captured_at":{"type":"string","title":"Captured At"},"amazon_price":{"type":"number","title":"Amazon Price"},"net_profit":{"type":"number","title":"Net Profit"},"roi_pct":{"type":"number","title":"Roi Pct"},"margin_pct":{"type":"number","title":"Margin Pct"},"referral_fee":{"type":"number","title":"Referral Fee"},"fba_fee":{"type":"number","title":"Fba Fee"},"flags":{"items":{"type":"string"},"type":"array","title":"Flags","default":[]},"modelled":{"type":"boolean","title":"Modelled","description":"True until real Keepa data is attached","default":true}},"type":"object","required":["retailer","retailer_slug","product_id","title","price","list_price","discount_pct","captured_at","amazon_price","net_profit","roi_pct","margin_pct","referral_fee","fba_fee"],"title":"Lead"},"LeadPage":{"properties":{"total":{"type":"integer","title":"Total"},"count":{"type":"integer","title":"Count"},"offset":{"type":"integer","title":"Offset"},"items":{"items":{"$ref":"#/components/schemas/Lead"},"type":"array","title":"Items"},"modelled":{"type":"boolean","title":"Modelled","default":true},"note":{"type":"string","title":"Note","default":"Amazon price is modelled as list_price x multiplier. Configure a Keepa API key for live pricing and rank history."}},"type":"object","required":["total","count","offset","items"],"title":"LeadPage"},"Retailer":{"properties":{"slug":{"type":"string","title":"Slug"},"name":{"type":"string","title":"Name"},"host":{"type":"string","title":"Host"},"platform":{"type":"string","title":"Platform"},"tier":{"type":"integer","title":"Tier","description":"1 open catalog .. 4 needs paid scraping"},"enabled":{"type":"integer","title":"Enabled"},"products":{"type":"integer","title":"Products"}},"type":"object","required":["slug","name","host","platform","tier","enabled","products"],"title":"Retailer"},"Sale":{"properties":{"retailer":{"type":"string","title":"Retailer"},"retailer_slug":{"type":"string","title":"Retailer Slug"},"product_id":{"type":"integer","title":"Product Id"},"title":{"type":"string","title":"Title"},"brand":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"pack_qty":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pack Qty"},"price":{"type":"number","title":"Price"},"list_price":{"type":"number","title":"List Price"},"discount_pct":{"type":"number","title":"Discount Pct"},"captured_at":{"type":"string","title":"Captured At"}},"type":"object","required":["retailer","retailer_slug","product_id","title","price","list_price","discount_pct","captured_at"],"title":"Sale"},"SalePage":{"properties":{"total":{"type":"integer","title":"Total"},"count":{"type":"integer","title":"Count"},"offset":{"type":"integer","title":"Offset"},"items":{"items":{"$ref":"#/components/schemas/Sale"},"type":"array","title":"Items"}},"type":"object","required":["total","count","offset","items"],"title":"SalePage"},"Stats":{"properties":{"retailers":{"type":"integer","title":"Retailers"},"products":{"type":"integer","title":"Products"},"price_snapshots":{"type":"integer","title":"Price Snapshots"},"amazon_products":{"type":"integer","title":"Amazon Products"},"matches":{"type":"integer","title":"Matches"},"last_scan":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Scan"}},"type":"object","required":["retailers","products","price_snapshots","amazon_products","matches","last_scan"],"title":"Stats"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}